+++ R/monitor_map.R
@@ -4,7 +4,6 @@
 #' @title Create Map of Monitoring Stations
 #' @param ws_monitor data list of class \code{ws_monitor}
 #' @param slice either a time index or a function used to collapse the time axis -- defautls to \code{get('max')}
-#' @param AQIStyle AQI breaks and colors to use ('1_3'|'8'|'24') -- defaults to '1_3'
 #' @param breaks set of breaks used to assign colors or a single integer used to provide quantile based breaks
 #' @param paletteFunc a palette generating function as returned by \code{colorRampPalette}
 #' @param add logical specifying whether to add points to an existing map
@@ -20,17 +19,10 @@
 #' 
 #' If \code{slice} is a function (not a function name) it will be used with argument \code{na.rm=TRUE} to
 #' collapse the time dimension. Thus, user defined functions must accept \code{na.rm} as a parameter.
-#' @details The colors used for the monitors will be chosen based on the AQIStyle which should match the
-#' width of any rolling mean that was applied to the data. Note that the AQI colors used for 1-hr and 3-hr
-#' rolling means are the same.
+#' @details Using a single number for the \code{breaks} argument will cause the algorithm to use
+#' quantiles to determine breaks.
 #' 
-#' Both breaks and paletteFunc must both be specified in order to use them. 
-#' 
-#' Using a single number for the \code{breaks} argument will cause the algorithm to use
-#' quantiles to determine breaks. This is reasonable behavior when coloring by size, or
-#' fire radiative power. In the case of coloring by 'datetime' or some other
-#' attribute where you need linear breaks you will have to create those breaks
-#' manually and pass them into the breaks argument.
+#' You can use AQI colors and 24-hr, daily average breaks by specifying \code{breaks=NULL, paletteFunc=Null}.
 #' @examples
 #' \dontrun{
 #' airnow <- airnow_load(20150901, 20150930)
@@ -39,24 +31,17 @@
 #' title('Max PM2.5 Levels in September, 2015)
 #'}
 
-monitor_map <- function(ws_monitor, slice=get('max'), AQIStyle='1_3',
-                        breaks = NULL, paletteFunc=NULL,
+monitor_map <- function(ws_monitor, slice=get('max'),
+                        breaks=7,
+                        paletteFunc=colorRampPalette(c('gray90','antiquewhite','yellow','red','firebrick')),
                         showCounties=TRUE, showLegend=TRUE,
                         add=FALSE, ...) {
   
   lon <- ws_monitor$meta$longitude
   lat <- ws_monitor$meta$latitude
   
-  AQIStyle <- as.character(AQIStyle)
-  
-  # Sanity check
-  if ( !(AQIStyle  %in% c('1_3','8','24')) )
-    stop(paste0('AQIStyle "',AQIStyle,'" is not recognized. Please use "1_3", "8" or "24".'))
-  
-  # Classify AQI level based on statistic at each station
-  if (is.null(slice)) {
-    stop("Need to specify a slice")
-  } else if (class(slice) == "function") {
+  # Create the 'slice'
+  if (class(slice) == "function") {
     # NOTE:  Need as.matrix in case we only have a single monitor
     allMissingMask <- apply(as.matrix(ws_monitor$data[,-1]), 2, function(x) { all(is.na(x)) } )
     data <- as.matrix(ws_monitor$data[,-1])
@@ -77,23 +62,25 @@ monitor_map <- function(ws_monitor, slice=get('max'), AQIStyle='1_3',
   # ----- Figure out names for a legend and colors for each point ---- 
   
   # If the user didn't use custom breaks then use AQI names and colors
-  if ( is.null(breaks)) {
-    breaks <- AQI[[ paste0("breaks_", AQIStyle) ]]
-    names <- AQI$names
-    levelColors <- AQI$colors
+  if ( is.null(breaks) ) {
+    breaks <- AQI$breaks_24
+    legendColors <- AQI$colors
+    legendLabels <- AQI$names
+    legendTitle <- 'AQI Levels'
   } else {
-    # If the used a scalar for the breaks then use that many quantiles for breaks
+    # If they used a scalar for the breaks then use that many quantiles for breaks
     if ( length(breaks) == 1 ) {
       probs <- seq(0,1,length.out=(breaks+1))
       breaks <- quantile(pm25, probs=probs, na.rm=TRUE)
     }
     # For each break, use the lower number as the name in the legend.
-    names <- unname(head(breaks, -1))
     levelColors <- paletteFunc(length(breaks) - 1)
+    legendLabels <- paste(sprintf("%.1f",breaks[-length(breaks)]),'--',sprintf("%.1f",breaks[-1]))
+    legendTitle <- 'Custom Levels'
   }
   
   # Create levels and use them to create a color mask
-  levels <- .bincode(pm25, breaks, include.lowest = TRUE)  
+  levels <- .bincode(pm25, breaks, include.lowest=TRUE)  
   cols <- levelColors[levels]
   
   # ------------------------------- GRAPHICS --------------------------------------
@@ -138,5 +125,5 @@ monitor_map <- function(ws_monitor, slice=get('max'), AQIStyle='1_3',
   argsList$col <- 'black'
   do.call(points, argsList)
   
-  if (showLegend) legend("bottomright", col=rev(levelColors), legend=rev(names), pch=16, cex=.7)
+  if (showLegend) legend("bottomright", col=rev(levelColors), legend=rev(legendLabels), pch=16, cex=.7)
 }